+++ server/src/infra/repositories/communication.repository.ts
@@ -15,9 +15,9 @@ import {
   WebSocketServer,
 } from '@nestjs/websockets';
 import { Server, Socket } from 'socket.io';
-import { DecorateAll, ExecutionTimeHistogram } from '../infra.utils';
+import { Instrumentation } from '../instrumentation';
 
-@DecorateAll(ExecutionTimeHistogram())
+@Instrumentation()
 @WebSocketGateway({
   cors: true,
   path: '/api/socket.io',
